requirements.txt
dvc
dvc-gs
joblib
matplotlib
mlflow
numpy
# Exporter for Google Cloud
opentelemetry-exporter-gcp-trace
pandas
pytest
scikit-learn
tzdata